fix: skip service overrides by service ID
Remove override checks by service ID alone. Such checks don't make
sense when a service ID may be used by multiple entities, so that
lookup has to be removed. Instead, replace it with going directly to
checking for an integration key for the service ID and creating the
entity mapping.